Checkout works, updated login
CheckOutScreen can now actually check out a book.
- The users name and ID is recorded alongside the book name and book ID in checkedOutBooks.csv
- The date checked out and the due date are recorded as well.

I updated the login from using membersData.indexOf(username/password).
Reason being that it would just find the first spot a that a username or password is found in the csv. For instance, if someones username was '1', and a persons before them had a phone number starting with 1, it would record that phone number as the username.
The changes made now are:
- Only search for the username and password once we've reached the username column.
- Set a boolean to see if membersData[i] and membersData[i + 1] (the password column) match the user and pass.

This is just insures we are always checking in the username and password columns rather than accidentally checking in an id column/phone number column etc.