Reverted copies check for catalogue and changed CFOSU functionality
Lara made a good point that the user should still be able to pre-book if the books copies are at 0. Therefore, we still need to see it be available on the catalogue!

I changed the CreateFilesOnStartUp functionality so that it checks if EVERY file exists, rather than just the CSVFiles folder