Overdue And ReserveBook functionality.
If a user has an overdue book(s), the system will display a window with all of their overdue books that need to be returned along with their return date. This is done by checking the users ID as they log in and going through the checkedOutBooks file.

I display a fake message saying a $10 penalty will be applied to their account if they don't return it in time. Nothing actually happens with that.

On startup, the system will now check if a users reserved book has reached it's date for checkout and move it to the checkedOutBooks.csv file.